requirements.txt
numpy==1.26.0
pandas==2.1.4
sqlalchemy==2.0.25
streamlit==1.45.0
plotly==5.18.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
matplotlib==3.8.3
setuptools==69.5.1